+++ mpitests/test_apply.py
@@ -553,7 +553,9 @@ class TestApply:
         assert np.allclose(local_state_vector, local_expected_output_cpu, atol=tol, rtol=0)
 
 
-class TestSparseHamiltonianExpval:
+class TestSparseHamExpval:
+    """Tests sparse hamiltonian expectation values."""
+
     def test_sparse_hamiltonian_expectation(self, tol):
         comm = MPI.COMM_WORLD
         commSize = comm.Get_size()
@@ -571,9 +573,9 @@ class TestSparseHamiltonianExpval:
                 0.1 + 0.1j,
                 0.1 + 0.2j,
                 0.2 + 0.2j,
+                0.2 + 0.3j,
                 0.3 + 0.3j,
-                0.3 + 0.4j,
-                0.4 + 0.5j,
+                0.3 + 0.5j,
             ],
             dtype=np.complex128,
         )
@@ -582,15 +584,17 @@ class TestSparseHamiltonianExpval:
         comm.Scatter(state_vector, local_state_vector, root=0)
 
         dev_gpumpi = qml.device("lightning.gpu", wires=3, mpi=True, c_dtype=np.complex128)
+        dev_gpu = qml.device("lightning.gpu", wires=3, mpi=False, c_dtype=np.complex128)
 
         dev_gpumpi.syncH2D(local_state_vector)
+        dev_gpu.syncH2D(state_vector)
 
         H_sparse = qml.SparseHamiltonian(Hmat, wires=range(3))
 
         comm.Barrier()
 
         res = dev_gpumpi.expval(H_sparse)
-        expected = 1
+        expected = dev_gpu.expval(H_sparse)
 
         assert np.allclose(res, expected)
 
@@ -615,6 +619,7 @@ class TestExpval:
         obs = operation(wires)
         expval_single_wire_no_param(tol, obs)
 
+    @pytest.fixture(params=[np.complex64, np.complex128])
     @pytest.mark.parametrize(
         "obs",
         [
@@ -626,13 +631,13 @@ class TestExpval:
             qml.PauliZ(numQubits - 2) @ qml.PauliZ(numQubits - 1),
         ],
     )
-    def test_expval_multiple_obs(self, obs, tol):
+    def test_expval_multiple_obs(self, obs, request, tol):
         """Test expval with Hamiltonian"""
         num_wires = numQubits
         comm = MPI.COMM_WORLD
 
-        dev_cpu = qml.device("default.qubit", wires=num_wires, c_dtype=np.complex128)
-        dev_gpumpi = qml.device("lightning.gpu", wires=num_wires, mpi=True, c_dtype=np.complex128)
+        dev_cpu = qml.device("default.qubit", wires=num_wires, c_dtype=request.params)
+        dev_gpumpi = qml.device("lightning.gpu", wires=num_wires, mpi=True, c_dtype=request.params)
 
         def circuit():
             qml.RX(0.4, wires=[0])